frontend/app/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
altair==4.2.0
    # via streamlit
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
blinker==1.5
    # via streamlit
cachetools==5.2.0
    # via streamlit
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via streamlit
commonmark==0.9.1
    # via rich
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via
    #   ipython
    #   validators
entrypoints==0.4
    # via
    #   altair
    #   jupyter-client
executing==1.1.1
    # via stack-data
gitdb==4.0.9
    # via gitpython
gitpython==3.1.29
    # via streamlit
htbuilder==0.6.0
    # via st-annotated-text
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via streamlit
ipykernel==6.16.1
    # via
    #   ipywidgets
    #   pydeck
ipython==8.5.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.0.2
    # via pydeck
iteration-utilities==0.11.0
    # via htbuilder
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.16.0
    # via altair
jupyter-client==7.4.3
    # via ipykernel
jupyter-core==4.11.2
    # via jupyter-client
jupyterlab-widgets==3.0.3
    # via ipywidgets
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
numpy==1.23.4
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==21.3
    # via
    #   ipykernel
    #   streamlit
pandas==1.5.1
    # via
    #   altair
    #   streamlit
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.2.0
    # via streamlit
prompt-toolkit==3.0.31
    # via ipython
protobuf==3.20.3
    # via streamlit
psutil==5.9.3
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyarrow==9.0.0
    # via streamlit
pydeck==0.7.1
    # via streamlit
pygments==2.13.0
    # via
    #   ipython
    #   rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   pandas
    #   streamlit
pytz==2022.5
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyzmq==24.0.1
    # via
    #   ipykernel
    #   jupyter-client
requests==2.28.1
    # via streamlit
rich==12.6.0
    # via streamlit
semver==2.13.0
    # via streamlit
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
st-annotated-text==3.0.0
    # via -r requirements.in
stack-data==0.5.1
    # via ipython
streamlit==1.13.0
    # via -r requirements.in
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
traitlets==5.5.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   pydeck
typing-extensions==4.4.0
    # via streamlit
tzdata==2022.5
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
urllib3==1.26.12
    # via requests
validators==0.20.0
    # via streamlit
wcwidth==0.2.5
    # via prompt-toolkit
widgetsnbextension==4.0.3
    # via ipywidgets
zipp==3.9.0
    # via importlib-metadata
